api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]>=3.1,<4
python-dotenv==1.0.1
cryptography>=43.0,<44
telethon>=1.36,<2
httpx>=0.27,<1
